+++ src/main/java/id/ac/ui/cs/advprog/gatra/scoring/strategy/TierScoringStrategy.java
@@ -2,5 +2,5 @@ package id.ac.ui.cs.advprog.gatra.scoring.strategy;
 
 public interface TierScoringStrategy {
     String getTierName();
-    double calculateBaseScore(double totalPointsEarned, int totalMembers);
+    double calculateBaseScore(double totalPointsEarned, long totalMembers);
 }
